Update README.md
This is good writing and a good start to your project. You need to think through how to provide value beyond what currently exists and how your recommendation engine would actually work... i.e., what are you going to train it on? As we progress, I think we may need to scale back a bit, but we can discuss more in person.

95/100